Gracefully handle novel info failures.
Often, sources are broken and fail when fetching novel info.

Instead of crashing back to command line, allow the user to choose a different source without repeating the title entry / search / other steps.

Note: This is a quick proof of concept and needs someone more familiar with the code base to clean it up and support login.