Fix fanmtl.com / resolve_future add exceptions (#2601)
* Fix fanmtl.com

resolve_futures has bad default values (None) for desc and unit. If not defined by the caller, it will error downstream the callstack. This commit fixes only the fanmtl source, and does not touch that default behaviour. I will address this in another pull request.

* Add exceptions to taskman resolve_future

Add better exceptions for an obscure error that occurred because resolve_future's unit and desc parameters are optional. While I'm not sure about making them not optional (because of parameter ordering), this seems like an ok solution.

* In taskman resolve_futures change from Exceptions to default parameter values